Fix: db 약관동의 여부 ENUM 타입으로 수정
