+++ src/main/java/com/example/mdmggreal/member/entity/Member.java
@@ -3,7 +3,6 @@ package com.example.mdmggreal.member.entity;
 import com.example.mdmggreal.global.entity.BaseEntity;
 import com.example.mdmggreal.ingameinfo.type.Tier;
 import com.example.mdmggreal.member.dto.request.SignUpRequest;
-import com.example.mdmggreal.member.type.Agree;
 import com.example.mdmggreal.member.type.OAuthProvider;
 import com.example.mdmggreal.member.type.Role;
 import jakarta.persistence.*;
@@ -13,6 +12,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.ColumnDefault;
 
+import static com.example.mdmggreal.global.entity.type.BooleanEnum.FALSE;
+import static com.example.mdmggreal.global.entity.type.BooleanEnum.TRUE;
 import static com.example.mdmggreal.ingameinfo.type.Tier.UNRANK;
 import static com.example.mdmggreal.member.type.OAuthProvider.NAVER;
 import static com.example.mdmggreal.member.type.Role.USER;
@@ -79,7 +80,6 @@ public class Member extends BaseEntity {
      * 약관 동의 여부
      */
     @Embedded
-    @Column(nullable = false)
     private Agree agree;
 
     /**
@@ -111,10 +111,10 @@ public class Member extends BaseEntity {
                 .mobile(null)
                 .profileImage(request.getProfileImage())
                 .agree(Agree.builder()
-                        .agreeAge(request.getAgrees().isAgreeAge())
-                        .agreeTerms(request.getAgrees().isAgreeTerms())
-                        .agreePrivacy(request.getAgrees().isAgreePrivacy())
-                        .agreePromotion(request.getAgrees().isAgreePromotion())
+                        .agreeAge(request.getAgrees().isAgreeAge() ? TRUE : FALSE)
+                        .agreeTerms(request.getAgrees().isAgreeTerms() ? TRUE : FALSE)
+                        .agreePrivacy(request.getAgrees().isAgreePrivacy() ? TRUE : FALSE)
+                        .agreePromotion(request.getAgrees().isAgreePromotion() ? TRUE : FALSE)
                         .build())
                 .role(USER)
                 .tier(UNRANK)
