--- src/main/java/com/example/mdmggreal/member/type/Agree.java
@@ -1,19 +1,0 @@
-package com.example.mdmggreal.member.type;
-
-import jakarta.persistence.Embeddable;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Embeddable
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Getter
-public class Agree {
-    private boolean agreeAge;
-    private boolean agreeTerms;
-    private boolean agreePrivacy;
-    private boolean agreePromotion;
-}
