+++ src/main/java/com/example/mdmggreal/oauth/dto/SignInRequest.java
@@ -1,6 +1,5 @@
 package com.example.mdmggreal.oauth.dto;
 
-import com.example.mdmggreal.member.type.Agree;
 import lombok.Getter;
 
 @Getter
